Add options.getSpotlightConnectionUrl, move Spotlight integration to sentry core
