+++ sentry/src/main/java/io/sentry/internal/SpotlightIntegration.java
@@ -1,13 +1,12 @@
-package io.sentry.android.core;
+package io.sentry.internal;
 
 import static io.sentry.SentryLevel.DEBUG;
 import static io.sentry.SentryLevel.ERROR;
 
-import io.sentry.Hint;
-import io.sentry.IHub;
-import io.sentry.Integration;
-import io.sentry.SentryEnvelope;
-import io.sentry.SentryOptions;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.OutputStream;
@@ -17,34 +16,33 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.zip.GZIPOutputStream;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-public class SpotlightIntegration
-    implements Integration, SentryOptions.BeforeEnvelopeCallback, Closeable {
+import io.sentry.Hint;
+import io.sentry.IHub;
+import io.sentry.Integration;
+import io.sentry.SentryEnvelope;
+import io.sentry.SentryOptions;
+import io.sentry.util.Objects;
 
-  private @Nullable SentryOptions options;
+@ApiStatus.Internal
+public final class SpotlightIntegration
+  implements Integration, SentryOptions.BeforeEnvelopeCallback, Closeable {
 
-  // developer machine should be the same across emulators
-  // see https://developer.android.com/studio/run/emulator-networking.html
-  private final @NotNull String SPOTLIGHT_URL = "http://10.0.2.2:8969/stream";
+  private @NotNull SentryOptions options = SentryOptions.empty();
+
   private final @NotNull ExecutorService executorService = Executors.newSingleThreadExecutor();
 
   @Override
   public void register(@NotNull IHub hub, @NotNull SentryOptions options) {
     this.options = options;
 
-    // todo check if we're on the emulator
-    if (options.getBeforeEnvelopeCallback() == null) {
+    if (options.getBeforeEnvelopeCallback() == null && options.getSpotlightConnectionUrl() != null) {
       options.setBeforeEnvelopeCallback(this);
     }
   }
 
   @Override
-  public @Nullable SentryEnvelope execute(@NotNull SentryEnvelope envelope, @Nullable Hint hint) {
-    if (options == null) {
-      return envelope;
-    }
+  public @NotNull SentryEnvelope execute(@NotNull SentryEnvelope envelope, @Nullable Hint hint) {
     try {
       executorService.execute(() -> sendEnvelope(envelope));
     } catch (RejectedExecutionException ex) {
@@ -53,36 +51,35 @@ public class SpotlightIntegration
     return envelope;
   }
 
-  private void sendEnvelope(SentryEnvelope envelope) {
+  private void sendEnvelope(final @NotNull SentryEnvelope envelope) {
     try {
-      final HttpURLConnection connection = createConnection();
+      final @NotNull String spotlightConnectionUrl = Objects.requireNonNull(options.getSpotlightConnectionUrl(), "Spotlight URL can't be null");
+
+      final HttpURLConnection connection = createConnection(spotlightConnectionUrl);
       try (final OutputStream outputStream = connection.getOutputStream();
-          final GZIPOutputStream gzip = new GZIPOutputStream(outputStream)) {
+           final GZIPOutputStream gzip = new GZIPOutputStream(outputStream)) {
         options.getSerializer().serialize(envelope, gzip);
       } catch (Throwable e) {
-        options
-            .getLogger()
-            .log(
-                ERROR,
-                e,
-                "An exception occurred while submitting the envelope to the Sentry server.");
+        options.getLogger()
+          .log(
+            ERROR,
+            e,
+            "An exception occurred while submitting the envelope to the Sentry server.");
       } finally {
         final int responseCode = connection.getResponseCode();
         options.getLogger().log(DEBUG, "Envelope sent to spotlight: %d", responseCode);
-
         closeAndDisconnect(connection);
       }
-    } catch (IOException e) {
-      options
-          .getLogger()
-          .log(ERROR, e, "An exception occurred while creating the connection to spotlight.");
+    } catch (final Exception e) {
+      options.getLogger()
+        .log(ERROR, e, "An exception occurred while creating the connection to spotlight.");
     }
   }
 
-  private @NotNull HttpURLConnection createConnection() throws IOException {
+  private @NotNull HttpURLConnection createConnection(final @NotNull String url) throws Exception {
 
     final @NotNull HttpURLConnection connection =
-        (HttpURLConnection) URI.create(SPOTLIGHT_URL).toURL().openConnection();
+      (HttpURLConnection) URI.create(url).toURL().openConnection();
     connection.setRequestMethod("POST");
     connection.setDoOutput(true);
 
@@ -92,9 +89,6 @@ public class SpotlightIntegration
 
     // https://stackoverflow.com/questions/52726909/java-io-ioexception-unexpected-end-of-stream-on-connection/53089882
     connection.setRequestProperty("Connection", "close");
-
-    //    connection.setConnectTimeout(options.getConnectionTimeoutMillis());
-    //    connection.setReadTimeout(options.getReadTimeoutMillis());
 
     connection.connect();
     return connection;
